refactor: rename panel components/ subfolder to parts/
