+++ frontend/src/components/pet/interactions/panels/battle/parts/arena-slot.tsx
