+++ frontend/src/components/pet/interactions/panels/battle/parts/battle-overlay.tsx
