+++ frontend/src/components/pet/interactions/panels/battle/parts/battle-setup.tsx
