+++ frontend/src/components/pet/interactions/panels/battle/parts/fighter-picker-card.tsx
