+++ frontend/src/components/pet/interactions/panels/battle/parts/opponent-picker-card.tsx
